pyproject.toml
[tool.poetry]
name = "nonebot-plugin-usage-statistics"
version = "0.1.0"
description = ""
authors = ["ssttkkl <huang.wen.long@hotmail.com>"]
readme = "README.MD"
packages = [{include = "nonebot_plugin_usage_statistics"}]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0"
nonebot-plugin-session = "^0.1.0"
nonebot-plugin-datastore = "^1.1.1"


[tool.poetry.group.dev.dependencies]
nb-cli = "^1.2.4"
nonebot-plugin-datastore = {extras = ["cli"], version = "^1.1.1"}


[tool.poetry.group.adapters.dependencies]
nonebot-adapter-onebot = "^2.2.4"
nonebot-adapter-qqguild = "^0.2.5"
nonebot-adapter-kaiheila = "^0.2.12"
nonebot-adapter-console = "^0.4.0"
nonebot-adapter-telegram = "^0.1.0b14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.nonebot]